requirements.txt
# SavosBot Club - Telegram Bot Requirements
# Python dependencies for the Telegram bot

# Core Telegram Bot Library
python-telegram-bot==20.7

# HTTP Client for website communication
aiohttp==3.9.1

# Environment variables management
python-dotenv==1.0.0

# Date and time utilities
python-dateutil==2.8.2

# JSON handling (built-in, but useful for complex operations)
# json - built-in

# File system operations (built-in)
# pathlib - built-in

# Async support (built-in)
# asyncio - built-in

# Logging (built-in)
# logging - built-in

# Type hints (built-in)
# typing - built-in

# Optional: For better performance
uvloop==0.19.0; sys_platform != "win32"

# Optional: For development
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional: For code formatting
black==23.11.0
flake8==6.1.0

# Optional: For monitoring
psutil==5.9.6
